Took care of the warnings, except for anonymous namespace, which is intentional

+++ models/TorchXorPaper/include/TorchXor.hpp
@@ -55,7 +55,7 @@ class TorchXor final: public Model{
     std::vector<double> get_weights(void) override {
         torch::Tensor vals = torch::nn::utils::parameters_to_vector(net->parameters()).clone();
         std::vector<double> w(vals.sizes()[0]);
-        for(int i = 0; i < w.size(); i++) {
+        for(int i = 0; i < static_cast<int>(w.size()); i++) {
           w[i] = vals[i].item<double>();
         }
         return w;
